+++ src/entities/script.h
@@ -75,6 +75,7 @@ public:
     QUrl remoteFileUrl(QString fileName);
     static bool scriptFromRepositoryExists(QString identifier);
     void setSettingsVariablesJson(QString json);
+    QString getSettingsVariablesJson();
     QJsonObject getSettingsVariablesJsonObject();
     void setSettingsVariablesJson(QJsonObject jsonObject);
     QString getScriptDirPath();